refactor(capture): move bridge helper out of routers
